# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RSS v2",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_api_messages_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "发布时间下界，UTC 秒",
              "title": "Since"
            },
            "description": "发布时间下界，UTC 秒"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "发布时间上界，UTC 秒",
              "title": "Until"
            },
            "description": "发布时间上界，UTC 秒"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{message_id}": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Get Message",
        "operationId": "get_message_api_messages__message_id__get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{message_id}/related": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Related Messages",
        "description": "共享关键词的其他消息；没有关键词时返回空列表。",
        "operationId": "related_messages_api_messages__message_id__related_get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelatedMessageResponse"
                  },
                  "title": "Response Related Messages Api Messages  Message Id  Related Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{message_id}/translate": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Translate Message",
        "operationId": "translate_message_api_messages__message_id__translate_post",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages/{message_id}/enrich": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Enrich Message",
        "operationId": "enrich_message_api_messages__message_id__enrich_post",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "description": "查询任务列表。",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(queued|running|succeeded|failed)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Retry Task",
        "description": "重新排队失败任务。",
        "operationId": "retry_task_api_tasks__task_id__retry_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeywordResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "text",
          "kind"
        ],
        "title": "KeywordResponse",
        "description": "一条检索关键词；text 用于展示，kind 用于检索优先级与界面标注。"
      },
...
      "MessageDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/MessageVersionResponse"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "external_id",
          "created_at",
          "updated_at",
          "versions"
        ],
        "title": "MessageDetailResponse",
        "description": "消息详情响应。"
      },
      "MessageListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "MessageListResponse",
        "description": "消息列表响应。"
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          },
          "latest_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageVersionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "external_id",
          "updated_at",
          "latest_version"
        ],
        "title": "MessageResponse",
        "description": "消息列表里的一行：消息本体与它的最新版本。"
      },
      "MessageVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "collected_at": {
            "type": "integer",
            "title": "Collected At"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/TranslationResponse"
            },
            "type": "array",
            "title": "Translations"
          },
          "enrichments": {
            "items": {
              "$ref": "#/components/schemas/EnrichmentResponse"
            },
            "type": "array",
            "title": "Enrichments"
          },
          "translation_task": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "enrichment_task": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_id",
          "version_number",
          "title",
          "summary",
          "content",
          "url",
          "published_at",
          "collected_at",
          "language",
          "content_hash",
          "translations",
          "enrichments"
        ],
        "title": "MessageVersionResponse",
        "description": "消息版本响应。"
      },
...
      "RelatedMessageResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "collected_at": {
            "type": "integer",
            "title": "Collected At"
          },
          "shared": {
            "items": {
              "$ref": "#/components/schemas/KeywordResponse"
            },
            "type": "array",
            "title": "Shared"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "source_id",
          "version_id",
          "title",
          "published_at",
          "collected_at",
          "shared"
        ],
        "title": "RelatedMessageResponse",
        "description": "相关消息；shared 是双方共有的关键词，按实体优先排列。"
      },
...
      "TaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TaskListResponse",
        "description": "任务列表响应。"
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "lease_until": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Until"
          },
          "input_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Version Id"
          },
          "output_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Version Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "target_kind": {
            "type": "string",
            "title": "Target Kind",
            "default": ""
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_label": {
            "type": "string",
            "title": "Target Label",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_type",
          "idempotency_key",
          "status",
          "attempts",
          "lease_until",
          "input_version_id",
          "output_version_id",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "description": "任务状态响应。\n\nmodel / prompt_version 来自非敏感的任务快照；target_* 由任务列表的服务层解析\n（消息标题或来源名），嵌入在消息详情里的任务不填目标，因此给默认值。"
      },
...
      "TranslationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message_version_id": {
            "type": "string",
            "title": "Message Version Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "key_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Masked"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_version_id",
          "status",
          "title",
          "summary",
          "content",
          "provider_id",
          "key_masked",
          "model",
          "prompt_version",
          "task_id",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "TranslationResponse",
        "description": "翻译结果响应。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}